[server] Remove false alerts DIV logs in leaders
Currently drainers in leader replicas is validating the DIV info from
real-time topic instead of version topic; as a result, a false warning
alert log will be printed everytime when a new data segment is consumed.

This RB suppresses such logs if the DIV error is detected after
END_OF_PUSH_RECIEVED and server is trying to print logs for the very
first data message in a new segment.

TODO:
Refactor VeniceWriter as well as drainer service so that DIV info
of a produced message can be propagated to the drainer service.

RB=3333816
BUG=VENG-9260
G=voldemort-reviewers
R=jlliu,sumane,anaberez,yurzhou,nithakka,ssen,shuhliu,xnma,xyin,xhao,smaji,adietz,goliu,sdwu,lcli,zpolicze,fvillega,akozma,lelu
A=shuhliu