[server][test] Build mapping from kafka url to consumer after get assigned consumer from KafkaConsumerService.
Previouly, newly assigned consumer for this version topic is not added to the mapping of kafkaUrlToConsumerMap in StoreIngestionTask, then it relies on getConsumer(...) to be added.
This will cause slowness or failures for some test cases. We did not get noticed because we just make SharedConsumer as default recently.
This rb adds the newly assigned consumer into the mapping.

Bonous: Change back the timeout settings for the test cases.

RB=3340212
BUG=VENG-9494
G=voldemort-reviewers
R=mnhuang,lcli,sumane,akozma,goliu,jlliu,adietz,yurzhou,lelu,xyin,fvillega,xnma,nithakka,anaberez,shuhliu,smaji,ssen,sdwu,zpolicze
A=fvillega,lcli