[da-vinci-client][common][test] Create a separate thread pool for future versions in leader-follower helix ST
Today, rebalance triggered current version transition from OFFLINE to STANDBY is not async in order to avoid over-rebalance in Helix.
With a limited size of thread pool, new batch push can be blocked if all threads are already occupied by ST busy with the synchronized
transition.

This rb introduces a new thread pool for the future version state transitions, so the state transition for the new (future) versions
can go through, so rebalance on current version will not block the new version push anymore.

Add a integration test and to verify its correctness. Also ran the same test without this change and confirm that the test cannot pass.

RB=3346626
BUG=VENG-9518
G=voldemort-reviewers
R=goliu,sumane,yurzhou,lelu,ssen,anaberez,akozma,zpolicze,fvillega,xnma,adietz,lcli,jlliu,smaji,mnhuang,sdwu,xhao,nithakka,xyin
A=xhao