[da-vinci][server] Made the consumer's resume more robust
There was a bug where the StorageUtilizationManager's store change listener
could attempt to resume consumption on partitions but in native replication,
this would loop over all consumers, both local and remote, and inevitably,
only one of those would be subscribed, not both. In this case, we would call
resume on a consumer that doesn't have that partition subscribed, and Kafka
would throw an IllegalStateException.

The fix is to add a safeguard in our ApacheKafkaConsumer wrapper class, so
that the call is only delegated to the actual Kafka consumer when assigned.
There is already a similar safeguard for pause, and so now resume is treated
symmetrically.

A new unit test is added to ensure that this behavior works as expected for
both pause and resume.

Unrelated: added Closeable to our KafkaConsumerWrapper interface.

RB=3362783
BUG=VENG-9554
G=voldemort-reviewers
R=smaji,yurzhou,lelu,adietz,shuhliu,xyin,zpolicze,nithakka,jlliu,sumane,sdwu,xhao,akozma,lcli,goliu,anaberez,mnhuang,xnma,ssen
A=mnhuang