[server][da-vinci] Make sure status report adapter can handle amplification factor in all cases
This RB address race condition issues in amplification factor.

The changes are:
(1) Refactor ReportStatusAdapter so that report logic is simplified and is compatible with amp = 1 case. This is achieved by introducing an atomic counter and an atomic boolean array for each user partition, so report is only triggered when all subPartitions have received the reports. The logic will also apply to multi-thread reporting case so there won't be the case when status like END_OF_PUSH_RECEIVED or STARTED is reported after COMPLETED. Also, it unifies the path for amp = 1 and amp > 1 so our code can have better test coverage.
(2) Found another race condition in amplification factor: If TS message is received before longRunningTask check of LEADER role, those non-leader sub partitions won't be able to fall from LEADER to FOLLOWER in time. Especially when there is no message in the RT topic. In integration tests it will stuck in waitVersion() and eventually times out. (Because it won't set leader topic in PCS as is done in follower logic). We reuse the check logic in long running task check when processing TS. For non-leader-subPartition, we force it to fall to FOLLOWER and subscribe to local VT (This is logically correct and expected) before processing TS.

RB=3383844
BUG=VENG-9606
G=voldemort-reviewers
R=ssen,xnma,xhao,sdwu,smaji,lelu,goliu,adietz,akozma,sumane,fvillega,shuhliu,lcli,nithakka,anaberez,xyin,mnhuang,zpolicze,yurzhou
A=mnhuang