[controller] TopicReplicator cleanup and rename to RealTimeTopicSwitcher
The TopicReplicator was initially an abstract class, with two implementations,
one for Brooklin (which is now deprecated) and the other for Leader Follower.
Given that we have only the latter now, and that the sole purpose of the
abstraction was to aid in this transition, we can now collapse the abstraction
and its unique implementation into a single class. This also provides an
opportunity to clarify function and parameter names in various places.

Deleted the following dead configs:

- controller.enable.topic.replicator.ssl (relevant only to Brooklin)
- controller.enable.topic.replicator (code already acted as always true)

Also moved the RealTimeTopicSwitcher from the common module to the
controller module, and changed its package name to c.l.v.ingestion.control

Although the changes are sufficiently dense that the version control treats
RealTimeTopicSwitcher as a whole new file rather than a simple move, the
functionality is identical to before.

Unrelated:

- Tweaked the retry logic for topic creation, since it was sometimes timing
  out in tests.

Unrelated test changes:

- In a previous commit, I tweaked TestVeniceHelixAdminWithIsolatedEnvironment
  so that one of its tests which was clobbering the class' BeforeMethod setup
  would instead get its own dedicated class, where the setup can be done just
  once and properly, but I forgot to commit the new file, so it is included
  now in this commit. Also applied the same pattern to another similar test
  in the same file.
- Deflaked several unrelated tests, including:
  - TestParentControllerWithMultiDataCenter
  - PushStatusStoreTest
  - TestBackupVersionDatabaseOptimization

RB=3383690
BUG=VENG-9620
G=voldemort-reviewers
R=lelu,shuhliu,anaberez,lcli,goliu,akozma,adietz,mnhuang,zpolicze,xyin,sumane,ssen,yurzhou,nithakka,xnma,sdwu,xhao,smaji,jlliu
A=nithakka