[controller][test] Enable system store auto materialization for parent controller in integration tests
This RB enables Meta/PushStatus auto-materialization in integration tests. System stores will be created during store creation and it will be the default behavior for integration test setup.
The scope of this RB only includes tests that already on two layer cluster structure. For other tests, it will be enforced in a separate RB.

RB=3391000
BUG=VENG-9633
G=voldemort-reviewers
R=adietz,fvillega,lcli,shuhliu,yurzhou,xnma,akozma,xhao,mnhuang,lelu,smaji,zpolicze,sdwu,goliu,ssen,anaberez,sumane,xyin,nithakka
A=xyin