[venice-common] Fix server slow shutdown
Venice server cannot be shut down within the required time window and causes deployment (stop, restart) to fail.
The reason is that there are many (several hundreds) VeniceWriters in MetaStoreWriter to close, each of them
takes ~1 or 2 seconds to send the END_OF_SEGMENT control messages, and today this is done in a sequential manner.
So, in total, it requires more than 10 minutes to finish.

This rb changes the way that MetaStoreWriter to close the VeniceWriters in parallel and waits for all
close operations to finish before moving on. With this, EI test shows that it now takes about tens of seconds
(~36s as one example shown in the testing part) to close all writers, which is more than 10x better.

RB=3410308
BUG=VENG-9634
G=voldemort-reviewers
R=yurzhou,ssen,jlliu,lelu,adietz,xnma,xhao,akozma,xyin,zpolicze,anaberez,sumane,goliu,lcli,nithakka,sdwu,smaji,fvillega,mnhuang
A=goliu,mnhuang