[server] Followers do not throw exception when processing records from unsubscribed RT
Problem: When a leader is consuming real-time topic and then demoting to follower,
even though ingestion task thread lets the consumer unsubscribe from real-time
topic and changes the state from leader to standby, shared consumer thread might
still have remaining RT records to be processed. As the state is changed to standby,
shouldProcessRecord function will throw exception and fail the whole ingestion task.

Fix: When a follower processes a record from RT, if it does not subscribe to RT anymore,
just log a message and skip the record. Otherwise, if the follower still subscribes to
RT, throw an exception.

RB=3432407
BUG=VENG-9656
G=voldemort-reviewers
R=yurzhou,fvillega,ssen,sumane,lelu,goliu,adietz,anaberez,xhao,mnhuang,xnma,xyin,lcli,sdwu,zpolicze,akozma,jlliu,nithakka,smaji
A=mnhuang