[common] Add base Storage Persona classes for Next Generation Quota System implementation to interface with Zookeeper.
Previously, each quota managed a single store, and vice versa.  This RB includes classes that will allow an additional type of metadata, a Persona, to be stored into Zookeeper.  Each Persona will contain a list of stores that are maintained under a single quota.  For now, only the Storage Accessor for Zookeeper is implemented.

RB=3441780
BUG=VENG-9678
G=voldemort-reviewers
A=jlliu