[controller][test] Enable system store auto materialization for parent controller in integration tests
This reverts commit 7f38e11ec0c14024b51d12cea6b5d2e1ab0ea2c3.

This RB enables Meta/PushStatus auto-materialization in integration tests. System stores will be created during store creation and it will be the default behavior for integration test setup.
The scope of this RB only includes tests that already on two layer cluster structure. For other tests, it will be enforced in a separate RB.

Major changes in this RB:
(1) Add two auto-materialization configs in parent controlller creation wrapper. We can still disable them (if we have good reason to do so) by overriding the configs.
(2) Adjust system store deletion order. If both system stores are on, we need to delete push status store first, as it will try to update meta system store when it is deleted. The reverted order will cause error and result in fail to delete it.
(3) Set cluster number to 1 and separate some tests from TestMultiDataCenterPush to new class TestMultiDataCenterAdminOperations as
    1. Current router won't be able to separate requests to different clsuter. In my testing, it results in each router thread contains differnt clsuter knowledge and keeps failing if we don't hit the correct router thread.
    2. Some tests pollutes admin channel and it doesn't seem to fix itself and for now I don't plan to fix it in this RB, so I move them out as they are not testing the push but some admin channel operations.
(4) Adjust MetaSystemStoreTest class to remove unnecessary empty push.

Not fixed in this RB:
(5) It seems like there are still a test failing after enable auto-materialization: PushStatusStoreTest suppress meta system store auto materialization
This might be related to test logics or hidden code bugs and for now I have disabled them, otherwise the change could be too big for review.
For other tests that comments out auto materialization is due to test logic setup.

RB=3439968
G=voldemort-reviewers
R=lcli,lelu,adietz,smaji,xyin,yurzhou,sdwu,akozma,zpolicze,nithakka,xnma,ssen,mnhuang,sumane,fvillega,goliu,shuhliu,anaberez,xhao
A=zpolicze