[server][da-vinci] Refactor AmplificationFactorAdapter and fix race conditions with Ingestion Isolation
This RB refactors both AmplificationFactorAdapter and StatusReportAdapter to make it easier to understand.
Now, StatusReportAdapter will include AmplificationFactorAdapter as amp factor handler, not the other way around.
StatusReportAdapter will include a nested class PartitionReportStatus for each user partition for better abstraction instead of 2-3 layers of map structures.
AmplificationFactorAdapter now includes a few util method that takes in Consumer / Predicate that will be applied on each subPartition(or its PCS) of a specific user partition.

This RB also fixes a rare race condition - ConcurrentModificationException on OffsetRecord.toBytes() when AMP + II is enabled.
This is caused by concurrent serialization and modification on a same subPartition's OffsetRecord in differnt drainer threads. Now it will actively serailize subPartition's OffsetRecord when it reports COMPLETED.

The rest of the code changes in this RB is just to adapt to the core changes. It appears that there are some auto-reformat happening in StoreIngestionTask that is done automatically, please feel free to ignore.

Note: During the debugging phase I found that: HELIX_OFFLINE_PUSH_ENABLED is not enabled for most of the test while it is enabled in PROD. I tried to enable it but it fails some other test. Let's revisit this config and try to enable it in the test for all tests by default.

RB=3516980
BUG=VENG-9791
G=voldemort-reviewers
R=lcli,nithakka,anaberez,sdwu,ssen,akozma,zpolicze,xnma,lelu,goliu,smaji,xhao,mnhuang,yurzhou,sumane,xyin,fvillega,adietz,shuhliu
A=zpolicze