[client] Rename StoreSchemaFetcher#getLatestWriteOperationSchema to StoreSchemaFetcher#getLatestUpdateSchema
A decision was made to separate Delete semantics from Write Compute so
that Write Compute only includes partial update. That means Write
Compute schema is only partial update schema. So, renaming this API to
reflect that the schema returned is used to create a (partial) update
request.

RB=3518976
BUG=VENG-9800
G=voldemort-reviewers
R=xyin,nithakka,yurzhou,ssen,zpolicze,sdwu,goliu,shuhliu,lelu,xhao,jlliu,adietz,mnhuang,fvillega,smaji,xnma,akozma,sumane,anaberez
A=goliu,jlliu