[compat][controller] Remove Delete semantics from Write Compute schema generation
Deployment order: this change should be deployed to child controllers
first. Ideally, it should be firstly deployed to the child controller
instance that is going to be the leader child controller for
test-cluster cluster in EI and venice-0 in PROD.

Due to the decision to remove Delete semantics from Write Compute, this
change makes the controller generate Write Compute schema is not a union
schema contains a Delete record schema. Instead, the generated Write
Compute schema contains only an Update record schema.
Note that the write path is still capable of handling Write Compute
Delete. We should remove that from the write path after we verify that
no user is or will be sending Write Compute Delete requests to Venice.

RB=3530677
BUG=VENG-9801
G=voldemort-reviewers
R=xnma,xyin,adietz,anaberez,shuhliu,zpolicze,goliu,ssen,jlliu,mnhuang,xhao,akozma,smaji,sumane,nithakka,fvillega,sdwu,yurzhou,lelu
A=jlliu