[thin-client][davinci][Router] Switched the underlying impl of batch-get/compute to streaming
This code change will enable Streaming in Router by default and removed the config of controlling
streaming feature, and it also switches to streaming for the blocking batch-get/compute APIs by default,
and remove the blocking impl. This code change doesn't clean up the blocking API support in Router
since the existing client versions will continue to use it and the blocking API usage will be gradually
drained with the adoption of the new release of client.
On the client side, the metrics of batch-get/compute API will fall into `streaming` category, which may
cause some confusion on client side if the customers are actively monitoring these metrics.

TODO:
Add retry support for compute request

RB=3530789
BUG=VENG-9551
G=voldemort-reviewers
R=xhao,mnhuang,adietz,anaberez,fvillega,nithakka,lelu,xnma,smaji,sumane,akozma,ssen,zpolicze,lcli,jlliu,sdwu,yurzhou,xyin,shuhliu
A=nithakka