[da-vinci][server] StoreIngestionStats refactoring
The StoreIngestionTask used to update some of the metrics only at each
1000 messages. This now happens on every message, as the periodic
accounting created an edge case where the future leader stalled metric
could have false positives. The reason for this is that when a store
has a very low write rate, stats would be updated too infrequently. As
a result, the leader stalled metric was operating on stale information
and occasionally thought that no data was ingested, when in fact it
was just that no stats were recorded. As part of this some of the
state in PartitionConsumptionState can be deleted. The schema of
AdminResponseRecord is also evolved to remove the corresponing state.

To counteract the added overhead of recording 7 extra metrics on each
message rather than every 1000 messages, the StoreIngestionStats is
refactored so that the per-store object instance is looked up once
during SIT construction rather than on every metric recorded. This
should reduce the number of map lookups per message by ~12, plus 2
more for Write Compute and 3 more for Active/Active. A similar change
could also be done for AggVersionedStorageIngestionStats and
AggVersionedDIVStats but that is left out of scope of the current
commit.

RB=3573220
BUG=VENG-9811
G=voldemort-reviewers
R=anaberez,xyin,jlliu,akozma,nithakka,smaji,mnhuang,zpolicze,sumane,shuhliu,xhao,goliu,lcli,lelu,adchen
A=smaji