[da-vinci-client] Fix a race in StoreIngestionTask.processConsumerActions
In function processConsumerActions(), it tries to peek() an action from the head of consumerActionsQueue, process it,
and then removes one from the head. In normal case it would work, but it is also possible that the head of queue changes
after the peek() and thus the action later gets removed is not the same one processed.

The reason the above race condition can happen is because consumerActionsQueue is a priority queue and its head position can
be superseded by another action with higher priority (e.g. KILL defined in ConsumerActionType).

The fix to the problem is fairly straightforward by using the PriorityBlockingQueue.remove to remove the action processed
instead of the head.

RB=3597931
G=voldemort-reviewers
R=shuhliu,xhao,xyin,zpolicze,adchen,mnhuang,mnagaraj,sumane,anaberez,goliu,lcli,fvillega,smaji,akozma,nithakka,jlliu
A=sumane