[h2v] Report DATASET_CHANGED when MR fails and dataset changes mid push
Today if user removes a file when mapper starts, MapReduce will fail. We should classify
this issue as a user error. Otherwise, the availability in H2V report will decrease.

When a file is removed/added in a dir, the dir modification time will change. We could
leverage this timestamp to check file removal/addition. When push job starts, capture the
modification time of input dir. When MR fails, capture the modification time again. If it
changes, report a new checkpoint DATASET_CHANGED.

RB=3575852
BUG=VENG-9855
G=voldemort-reviewers
R=fvillega,nithakka,anaberez,adchen,lelu,zpolicze,xhao,sumane,smaji,lcli,mnhuang,xyin,goliu,jlliu,akozma
A=mnagaraj,nithakka