[server][da-vinci][common] Remove unnecessary DIV metrics for per store system stores
This RB is the first phase of system store metrics cleanup.
It focuses on removing DIV latency stats for per user store system stores from being reported to InGraphs as planned in the design.

After this change:
(1) No latency related metrics will be registered and reported to InGraphs
(2) No letency related metrics will be recorded.
(3) Due to limited scope, DIVStats is still unaware of store name and for now it will keep creating all latency sensors locally. It is considered fine as (1)/(2) is more important. We can revisit this to see if we want to make DIVStats store-name aware after more phases of metrics cleanup.

Other things related to this change:
(1) Create a LatencySensor class to formulate latency sensor creation, recording and reporting to make codebase looks cleaner.
(2) Remove MIN latency metrics from DIV classes as discussed.
(3) Rename DIVStatsCounter to DIVStatsGauge as it is not a counter and the original name is very confusing.

RB=3587174
G=voldemort-reviewers
R=lcli,sumane,adchen,smaji,lelu,shuhliu,xhao,goliu,nithakka,fvillega,akozma,zpolicze,anaberez,mnhuang,xyin
A=smaji