[router][common][test] Make sure router always use customized view based routing repository
This RB makes main code and tests use CV based routing repo for Router as it has been rolled out for PROD for long time. This is critical as tests should reflect the PROD status to make sure features are working properly.

Discovered the fact that STREAM_REPROCESSING is not working as current implementation is not supported in the CV based routing repository. Given the fact that this feature is not enabled, this RB commentted out the related tests and will fix it in the following RB.

Also, this RB removes `OnlineInstanceFinderDelegator` and `PartitionStatusOnlineInstanceFinder` as we will now support only one mode for Router.

RB=3608346
BUG=VENG-9965
G=voldemort-reviewers
R=anaberez,mnagaraj,xyin,akozma,mnhuang,smaji,goliu,adchen,lelu,shuhliu,xhao,lcli,sumane,zpolicze,fvillega,nithakka
A=fvillega