[da-vinci][server] Support server and da-vinci-client per store metrics unregisteration upon store deletion
When config unregister.metric.for.deleted.store.enabled is enabled,
server DIV, storage engine, ingestion and http request metrics will
be unregistered when a store is deleted.

When da-vinci-client detects that a store does not exist in venice
and deletes the store locally, DIV, storage engine, ingestion and
rocksdb memory usage metrics will be unregistered if the config is
enabled.

RB=3607600
BUG=VENG-9931
G=voldemort-reviewers
R=lcli,sumane,fvillega,jlliu,adchen,anaberez,goliu,xhao,smaji,xyin,lelu,zpolicze,akozma,nithakka,mnhuang,mnagaraj
A=goliu