[router][common] Only check resource availbility before checking offline resources for new current version in router
This RB checks resource availability in RoutingDataRepository before checking offline partitions in VeniceVersionFinder in order not throwing VeniceNoHelixResourceException for version not found issue when version resource assignment is not yet updated in the Customized View based routing repository. Add a new API in the OnlineInstanceFinder to check the resource existence.

RB=3607924
BUG=VENG-9918
G=voldemort-reviewers
R=smaji,lelu,xyin,fvillega,mnhuang,adchen,lcli,xhao,zpolicze,nithakka,shuhliu,mnagaraj,akozma,goliu,anaberez,sumane
A=lelu,xyin