[server][common][controller][tool] Remove incremental push enabled checks from venice-server code
Since IncPush2RT is the only supported behavior, all inc push enabled stores
will have hybrid configs. Hence we can solely rely on isHybrid() in lieu of
isIncrementalPushEnabled check.
Also, removed the support for setting target version for an incremental push
as it is not in use.
Deleted a config: "target.version.for.incremental.push"

RB=3613354
G=voldemort-reviewers
R=jlliu,nithakka,akozma,amsaxena,zpolicze,goliu,lelu,mnagaraj,mnhuang,xhao,anaberez,shuhliu,smaji,fvillega,adchen,xyin
A=jlliu