[server] Fix exception handlers in ParticipantStoreConsumptionTask to prevent its silent death
When an exception is caught in ParticipantStoreConsumptionTask::run, before logging
it we first check if it is a redundant exception using RedundantExceptionFilter::
isRedundantException(e.getMessage()). When e.getMessage() returns null,
isRedundantException throws NPE. This NPE gets caught in the outer catch block
but only to end in the same fate. Whenever this happens ParticipantStoreConsumptionTask
dies silently without logging any exit statements.

This RB fixes this issue and also added safeguards at several places to prevent
NPE due to the use of e.getMessage().

RB=3613255
BUG=VENG-9972
G=voldemort-reviewers
R=adchen,fvillega,jlliu,smaji,lelu,amsaxena,anaberez,xhao,goliu,mnagaraj,akozma,mnhuang,zpolicze,nithakka,xyin,shuhliu
A=smaji,goliu