[venice-fast-client] Added a default implementation of R2 Client in venice-fast-client
This code change adds a HttpClient5 based R2 Client in venice-fast-client,
which is being used by LinkedIn.

This RB won't automatically create such instance during fast-client construction, and
it still requires user to provide a R2 Client instance in ClientConfig.
So the procedure will be:
1. Construct a HttpClient5 based R2 client in the application code.
2. Pass it to a ClientConfig.
3. Construct a Venice-fast-client instance by the ClientConfig.

This methodology will apply to the dependencies, such as venice-thin-client/DaVinciClient
required by Venice-fast-client.

RB=3616551
BUG=VENG-9977
R=jlliu,mnagaraj,xyin,lelu,lcli,akozma,fvillega,sumane,adchen,nithakka,xhao,smaji,shuhliu,mnhuang,zpolicze,anaberez
A=fvillega