[da-vinci] Fix bug in object cache version clean up which triggered NPE
Test should have caught this, but the test itself had a bug which enabled this
bug. The versionTopicToCacheEngineMap in ObjectCacheBackend is keyed to version
topic names, NOT storeName. This can trigger an NPE on store change and not
clean things up on version swap.  The test in turn, had provided to the Mock
the incorrect storeName (it instead gave the version topic name as the store
name).  This made it so the test wouldn't catch the bug.  So fixed the test
to trigger the bug, and then fixed the bug.

RB=3621555
G=voldemort-reviewers
R=xyin,goliu,xhao,lcli,mnhuang,jlliu,sumane,nithakka,lelu,mnagaraj,fvillega,adchen,shuhliu,anaberez,smaji,akozma
A=sumane