[server][controller][admin-tool] WIP DONT REVIEW Schema updates for change capture events and store version config
We add in the concept of a 'view' array to the store version config.  The view is meant to define the parameters for a parallel set of topics which a leader will emit events to.  To start with, we'll only have one view called a 'changeCapture' view (name TBD).  This informs us to create a topic which will receive events as we ingest on leader topics which conform to the schema described in RecordChangeEvent.avsc.

RB=3625791