[server] Use interpolation for logging in venice-common (#4)
<!--
Add a list of affected components in the PR title in the following
format:
[component1]...[componentN] Concise commit message

Valid component tags are: [da-vinci], [server], [controller],
[router], [samza], [vpj], [fast-client], [thin-client], [alpini],
[admin-tool], [test], [build], [doc], [script]

Example title: [server][da-vinci] Use dedicated thread to persist data
to storage engine

Note: PRs with titles not following the format will not be merged
-->
This rb mainly contains the logging clean ups for venice-common module.
<!--
Describe
- What changes to make and why you are making these changes.
- How are you going to achieve your goal.
- Describe what testings you have done, for example, performance testing
etc.

If this PR resolves an issue be sure to include "Resolves #XXX" to
correctly link and close the issue upon merge.
-->

## How was this PR tested?
<!--
If you're unsure about what to test, where to add tests, or how to run
tests, please feel free to ask. We'd be happy to help.
-->
Internal CI (yellow)

## Does this PR introduce any user-facing changes?
<!--
If yes, please clarify the previous behavior and the change this PR
proposes - provide the console output, description and/or an example to
show the behavior difference if possible.
If no, choose 'No'.
-->
- [x] No. You can skip the rest of this section.
- [ ] Yes. Make sure to explain your proposed changes and call out the
behavior change.