[h2v] Repush optimization for chunked topics
Implemented object reuse for chunked topics. Because the logic is more
complex, it is still not as efficient as non-chunked, but nonetheless
better than before. The memory requirement is now to buffer the bytes
of all shuffled chunks rather than the objects of all shuffled events.

Also removed an unnecessary optional for the reducer hot path for all
implementations (including regular VPJ and repush).