[vpj][doc] Repush optimization for chunked topics (#27)
Implemented object reuse for chunked topics. Because the logic is more
complex, it is still not as efficient as non-chunked, but nonetheless
better than before. The memory requirement is now to buffer the 
serialized bytes of all shuffled chunks rather than the deserialized 
objects of all shuffled events.

Also removed an unnecessary optional for the reducer hot path for all
implementations (including regular VPJ and Repush).

Created ops_guide and user_guide directories in the docs, and started
populating them with pages on Repush and TTL.