[thin-client] Changed the retry strategy of store init
In thin-client, if the store init isn't done in the start method because
of the delay start of d2 client, each request could trigger the
store init procedure at runtime, and this has introduced some issue
in the past especailly when the store isn't in a good state, such
as store was deleted/disabled and the 10-time retry triggered by
each request will put the application thread pool in a stuck state,
and application's fallback logic can't kick in since the Venice call
will be blocked by the internal store init procedure.

This code change introduces a change of the retry strategy:
If d2 client isn't ready during the start invocation, the store init
with retry triggered by the online request will only be executed once,
and if the store init fails, all the subsequent requests will receive
an exception immediately.

Here are the reasons:
1. Increase the store init success rate at the startup time especially
   in test env.
2. Don't stuck the application threads when the store isn't in a good state,
   so that application fallback logic could kick in.

This code change also fixed a flaky test in AvroGenericStoreClientImplTest.