Allow RMD retrieval and parsing in TTL filter at Mapper stage
This commit introduced some refactoring changes and the implementation of for RMD retrieval and parsing in TTL filter.

RMD retrieval and parsing:
When repush TTL function is enabled, the VPJ will make a call onto controller to retrieve all RMD schemas belonging to this store on a temp folder on HDFS.
In the Mapper stage, when ttl filter is inited, the schems will be fetched to further parse the RMD records. Based on the TTL config, the record will be either ignored or filtered.

Refactoring:
1. Move some Controller/SSL related methods from VPJ main class to an Utils class.
2. Move some Rmd related methods from da-vinci package to the common page as a new Utils class.