[specs] Add specs subdirectory and spec for LeaderFollowerModel
This spec is meant to define the abstract flow of Venice leader follower
model. The intention is that this spec can be used as a starting point
for refinement as we add enhancements to deal with corner cases in
the Venice leader election or write path.

TODO: Version push is conspicuously missing.  Should add that as a
seperate state possibiliity (or leave it to later refinements).

Some things that could be added with this spec:

Deletes (thats a fun one!)
Write Compute Operations
A/A
Split Brain