[admin-tool] Fixed the admin tool command to dump admin topic
There was a bug in AdminTool#printObject method, which can't handle
a list of response because of the internal filtering logic.
Since field filtering is not being used actively, this code change
removes the field filtering option.
In the meantime, this code change also skips the Control Message
when dumping admin topic to avoid confusion.