[specs] More cleanup and remove version push from model
Gaojie pointed out that version push in this model maybe doesn't make sense since
we specify a leader follower state per version (instead of across all versions).
So It might make more sense to make a higher level spec that uses the LeaderFollower
as objects assigned to each version state.  So we'll use that in a future spec.

Also pulled MAX_WRITES variable into the AbstractVeniceLeaderFollower since it's not
needed in AbstractVeniceLeaderFollower