[controller] Handled execution id gap when firstly becoming a leader controller
Before this code change, `AdminConsumptionTask` already has logic to handle
execution id gap when producers are different and the missing scenario is that
the execution id gap could happen in the first poll request where there is no
previous producer info in memory.
This PR seeks to one older offset than the persisted offset to initialize
the in memory producer info.