[da-vinci][server] Meta system store waits for the fully consumption of the localVT before switching to RT
Today, for a system meta store, when a standby partition is in the process of switching to the leader,
it needs to check the time difference between now and the last local VT message timestamp and proceeds
to finish switching when the difference is larger than PROMOTION_TO_LEADER_REPLICA_DELAY_SECONDS setting.

One intention of having such delay is to allow the standby to fully consume its own local VT messages first.
However, in some cases, the delay can not gurantee and this turns out to be an issue when the leader RT topic
in kafka is truncated (and becomes empty) due to retention settings. When RT is empty and thus no message to
consume when the switch happens, the offset lag between the latest RT and the periodically persisted OffsetRecord
could not be mitigated anymore (if not fully consumed) and cause state transition to fail, thus error leader
replica.

This change fixes the issue by let the switch to wait until all local VT messages in meta system store to
be fully consumed. After the fully consumption, there should be no gap between the latest RT and the locally
persisted offset by design, thus the state transition can success.