[da-vinci][test] Add a new metric to capture offset rewind in Kafka for user stores
In case of hitting a Kafka storage problem, one issue we have seen in the past is that Kafka
partition on a consumer can lose its offset and start re-consuming the entire partition from
the beginning. We want to detect this offset drop/rewind event and report when it happens.
One way to detect it from the Venice servers's point is to compare its local persisted offset
for that partition with the end offset in Kafka and report when the persisted one is greater.
The comparison is only done at the time when server subsscribes to that partition.

This rb also introduces a new version level metrics to report the offset rewind (backup,
current, future, and total) for each user store and it records the occurrence number for
the rewind events across all partitions for the store versions. A unit test is also added
to verity the behaviour of the metrics is as expected.