[test] Use lsof without abs path IsolatedIngestionServerTest::testReleaseTargetPortBinding (#98)
Both RHEL & CentOS require an absolute path of lsof. However, this path is not the same 
on Ubuntu. Also, the abs path is not required on ubuntu. So we first try with the abs path
and if it returns an empty string, that likely means that the command did not succeed and
hence we try it again without the abs path.