[da-vinci][test] Add a new metric to capture offset rewind in Kafka for user stores (#99)
* [da-vinci][test] Add a new metric to capture offset rewind in Kafka for user stores

In a Kafka problem hit previous, if durability related configs are not set correctly, Kafka could lose data and reset end-offset of the impacted partitions to smaller offsets; new writes on the impacted partitions will have old(small) offsets than the checkpoint offsets in Venice, and thus Venice drops and loses new writes. (Another side effect is that Venice servers could re-consume the entire partition from the beginning). We want to detect this end-offset drop/rewind event and report when it happens. One way to detect it from the Venice servers's point is to compare its local persisted offset for that partition with the end-offset in Kafka and report when the persisted one is greater. The comparison is only done at the time when server subscribes to that partition.

This rb also introduces a new version level metrics to report the offset rewind (backup, current, future, and total) for each user store and it records the occurrence number for the rewind events across all partitions for the store versions. A unit test is also added to verity the behaviour of the metrics is as expected.

Note that this rb only targets store's version topics and the realtime topic is handled by the existing logic e.g. LeaderFollowerStoreIngestionTask.checkAndHandleUpstreamOffsetRewind.