[server][controller][admin-tool][compat]  Schema updates for c… (#57)
* [server][controller][admin-tool] Schema updates for view configs to venice stores and store versions

We add in the concept of a 'view' map to the store version config.  The view is meant to define the parameters for a kind of store procedure that venice leaders will act on during ingestion.  To start with, we'll only have one view called a 'changeCapture' view.  This informs us to create a topic which will receive events as we ingest on leader topics which conform to the schema described in RecordChangeEvent.avsc.

Co-authored-by: Zac Policzer <zpolicze@zpolicze-mn1.linkedin.biz>