[controller] WIP Transmit versionSwap message when version changes
We change versions under one of three cases.  Push completion (so long as version swap isn't deferred)
version setting (makeCurrent) or rollback.  We'll only transmit the version swap message if one of the
versions we're transiting to is hybrid, and as a temporary measure, only transmit the message if the
store has a view set up for either version.  This is perhaps temporary, but this removes the blast radius
until we've made changes to make this all relevant for other venice stores.