[controller][test] Fix the error in getting largest used version number when recreating a per-user system store
When a store is deleted, put into the graveyard, and then recreated through a createStore API, it is expected to
assign a version number to the system stores that is 1 more than the value retrieved from the graveyard. Today,
for system stores, in function getStoreLargestUsedVersion, it checks if the regular store of such system stores
has been created and only query store graveyard when it hasn't. This might have an issue because in the createStore
workflow, regular store creation could happen before the query (or after it) if Admin execution task responds
quickly (or slowly) and when the regular store is created, it doesn't ask store graveyard and wrong version number
is returned.

This rb fixes the above issue by always querying the store graveyard for the largest used version number during
createStore workflow. It also adds condition waitings in the related tests to make it more stable.